+++ guicast/bctextbox.C
@@ -621,7 +621,7 @@ void BC_TextBox::deactivate()
 	draw();
 }
 
-void BC_TextBox::repeat_event(int64_t duration)
+void BC_TextBox::repeat_event(int duration)
 {
 	if(duration == top_level->get_resources()->tooltip_delay &&
 		tooltip_text[0] != 0 &&
